+++ dev/Network.js
@@ -90,13 +90,13 @@ class Network {
             }
         }
 
-        if(["uniform"].includes(this.weightsConfig.distribution)) {
+        if(this.weightsConfig.distribution == "uniform") {
             this.weightsConfig.limit = weightsConfig && weightsConfig.limit!=undefined ? weightsConfig.limit : 0.1
 
         } else if(this.weightsConfig.distribution == "gaussian") {
 
             this.weightsConfig.mean = weightsConfig.mean || 0
-            this.weightsConfig.stdDeviation = weightsConfig.stdDeviation || 1          
+            this.weightsConfig.stdDeviation = weightsConfig.stdDeviation || 0.05        
         }
 
         // Status
@@ -170,7 +170,9 @@ class Network {
         layer.adaptiveLR = this.adaptiveLR
         layer.activationConfig = this.activationConfig
         layer.dropout = this.dropout
-        layer.weightsConfig = this.weightsConfig
+
+        layer.weightsConfig = {}
+        Object.assign(layer.weightsConfig, this.weightsConfig)
         layer.weightsInitFn = NetMath[layer.weightsConfig.distribution]
 
         if(this.rho!=undefined) {
@@ -190,6 +192,7 @@ class Network {
         }
 
         if(layerIndex) {
+            layer.weightsConfig.fanIn = this.layers[layerIndex-1].size
             this.layers[layerIndex-1].assignNext(layer)
             layer.assignPrev(this.layers[layerIndex-1])
         }